refactor: change styles
