+++ frontend/tailwind.config.cjs
@@ -18,5 +18,5 @@ module.exports = {
       },
     },
   },
-  plugins: [require('@tailwindcss/forms')],
+  plugins: [],
 };
